PR-CLEANUP-5 — tool_handlers consolidation + _helpers rename
Two cleanups on the core/cli/tool_handlers/ surface, driven by the
new Naming CANNOTs landed in PR-DOCS-CANT-CAN.

(1) Fold six small handler files into one

The files were data.py / notification.py / output.py / offload.py /
computer_use.py / calendar.py — each <50 LOC, each wrapping exactly
one tool class via the same boilerplate (instantiate Tool, wrap
aexecute in a closure, return {name: handler}) with zero shared
state. The split was pure noise. All six builders now live in
core/cli/tool_handlers/single_tool.py (169 LOC, grouped by section
comments). The _build_<area>_handlers symbol names are preserved
verbatim — package __init__.py + external test callers continue to
import them unchanged, only the source module-name moves.

(2) Rename _helpers.py -> clarification.py

The Naming CANNOT row forbids _helpers / _utils filenames once a
caller appears. The file owned two functions — _clarify (builds the
clarification_needed follow-up-question response) and
_safe_delegate (turns missing-kwarg exceptions into _clarify
calls). Both are clarification-shaped responses for tool dispatch,
so "clarification" describes the file's actual role. 6 importers +
2 stale docstring references (core/tools/arxiv.py,
core/tools/seed_pool_search.py) updated.

Codex MCP verified (read-only): 8 PASS / 1 FAIL (intentional —
computer_use module docstring was lifted into the function
docstring rather than dropped, content preserved) / 2 FLAG
(addressed pre-commit: single_tool.py was untracked, now staged).
