+++ core/cli/tool_handlers/plan.py
@@ -12,7 +12,7 @@ from __future__ import annotations
 import logging
 from typing import Any
 
-from core.cli.tool_handlers._helpers import _clarify
+from core.cli.tool_handlers.clarification import _clarify
 from core.ui.console import console
 
 log = logging.getLogger(__name__)
