+++ core/cli/tool_handlers/task.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from typing import Any
 
-from core.cli.tool_handlers._helpers import _clarify
+from core.cli.tool_handlers.clarification import _clarify
 
 
 def _build_task_handlers() -> dict[str, Any]:
